Fix reviews API 500 caused by migration FK type mismatch.
Use INT UNSIGNED foreign keys for course_reviews and student_notifications, return empty reviews when table is missing, and log review load failures.

+++ api/migrations/00063_create_course_reviews_table.sql
@@ -3,8 +3,8 @@
 CREATE TABLE course_reviews (
     id BIGINT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
     tenant_id INT UNSIGNED NOT NULL,
-    course_id BIGINT UNSIGNED NOT NULL,
-    student_id BIGINT UNSIGNED NOT NULL,
+    course_id INT UNSIGNED NOT NULL,
+    student_id INT UNSIGNED NOT NULL,
     rating TINYINT UNSIGNED NOT NULL,
     comment TEXT NULL,
     tags JSON NULL,
